% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{app_areas}
\alias{app_areas}
\title{Data about application areas of selected measures}
\format{
A data.frame with 5 variables:
\describe{
\item{Subbasin}{Subbasins aggregating SWAT setups.}
\item{Setup_name}{SWAT setup name.}
\item{RCH}{Reach number correcting to reach or subbasin numbers in SWAT ouput files.}
\item{Measure}{Implemented measure code.}
\item{AREAkm2Measure}{Area in km2 where measures have been implemented.}
}
}
\usage{
app_areas
}
\description{
Data showing size of areas where measures have been applied in each small catchment.
}
\keyword{datasets}
