% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basins}
\alias{basins}
\title{Basin data}
\format{
A sf data.frame with 7 variables:
\describe{
\item{Setup_name}{SWAT setup name}
\item{Reach}{Reach number correcting to reach or subbasin numbers in SWAT ouput files.}
\item{CID}{Catchment ID. It is the same as in rivers CID.}
\item{Subbasin}{Subbasins aggregating SWAT setups.}
\item{Area}{Area in km2}
\item{CIDto}{CID of subbasin where water flows out of this subbasin}
\item{geometry}{Geometry of objects}
}
}
\usage{
basins
}
\description{
Data of basins used in modeling.
}
\keyword{datasets}
