% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_clean_data.R
\name{clean_swat_output}
\alias{clean_swat_output}
\title{Function to clean imported SWAT model output files}
\usage{
clean_swat_output(df)
}
\arguments{
\item{df}{Data.frame of imported output.*** SWAT file}
}
\value{
Data.frame of cleaned outoup files
}
\description{
This function removes unnecessary columns from imported SWAT output files.
}
\examples{
clean_swat_output(daily_sub)
}
