% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_averaging.R
\name{collapse_results_to_setups}
\alias{collapse_results_to_setups}
\title{Function to sum up output data over setups}
\usage{
collapse_results_to_setups(df)
}
\arguments{
\item{df}{Data.frame of imported output.*** SWAT file}
}
\value{
Data.frame of summed up data over setups
}
\description{
This function averages output data imported from SWAT output.rch
and output.sub files. Data should have been imported by using
extract_watersheds_output function.
}
