% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daily_rch}
\alias{daily_rch}
\title{Daily output.rch data}
\format{
A data frame with 55 variables:
\describe{
\item{date}{Date}
\item{Year}{Year}
\item{Month}{Month}
\item{Day}{Day}
\item{FILE}{File type}
\item{RCH}{Reach number.}
\item{MON}{Day}
\item{AREAkm2}{Area drained by reach (km2).}
\item{FLOW_INcms}{Average daily streamflow into reach during time step (m3/s).}
\item{FLOW_OUTcms}{Average daily streamflow out of reach during time step (m3/s).}
\item{EVAPcms}{Average daily rate of water loss from reach by evaporation during time step (m3/s).}
\item{TLOSScms}{Average daily rate of water loss from reach by transmission through the streambed during time step (m3/s).}
\item{SED_INtons}{Sediment transported with water into reach during time step (metric tons).}
\item{SED_OUTtons}{Sediment transported with water out of reach during time step (metric tons).}
\item{SEDCONCmg_kg}{Concentration of sediment in reach during time step (mg/L).}
\item{ORGN_INkg}{Organic nitrogen transported with water into reach during time step (kg N).}
\item{ORGN_OUTkg}{Organic nitrogen transported with water out of reach during time step (kg N).}
\item{ORGP_INkg}{Organic phosphorus transported with water into reach during time step (kg P).}
\item{ORGP_OUTkg}{Organic phosphorus transported with water out of reach during time step (kg P).}
\item{NO3_INkg}{Nitrate transported with water into reach during time step (kg N).}
\item{NO3_OUTkg}{Nitrate transported with water out of reach during time step (kg N).}
\item{NH4_INkg}{Ammonium transported with water into reach during time step (kg N).}
\item{NH4_OUTkg}{Ammonium transported with water out of reach during time step (kg N).}
\item{NO2_INkg}{Nitrite transported with water into reach during time step (kg N).}
\item{NO2_OUTkg}{Nitrite transported with water out of reach during time step (kg N).}
\item{MINP_INkg}{Mineral phosphorus transported with water into reach during time step (kg P).}
\item{MINP_OUTkg}{Mineral phosphorus transported with water out of reach during time step (kg P).}
\item{CHLA_INkg}{Algal biomass transported with water into reach during time step (kg chl-a).}
\item{CHLA_OUTkg}{Algal biomass transported with water out of reach during time step (kg chl-a).}
\item{CBOD_INkg}{Carbonaceous biochemical oxygen demand of material transported into reach during time step (kg O2).}
\item{CBOD_OUTkg}{Carbonaceous biochemical oxygen demand of material transported out of reach during time step (kg O2).}
\item{DISOX_INkg}{Amount of dissolved oxygen transported into reach during time step (kg O2).}
\item{DISOX_OUTkg}{Amount of dissolved oxygen transported out of reach during time step (kg O2).}
\item{SOLPST_INmg}{Soluble pesticide transported with water into reach during time step (mg active ingredient).}
\item{SOLPST_OUTmg}{Soluble pesticide transported with water out of reach during time step (mg active ingredient).}
\item{SORPST_INmg}{Pesticide sorbed to sediment transported with water into reach during time step (mg active ingredient).}
\item{SORPST_OUTmg}{Pesticide sorbed to sediment transported with water out of reach during time step (mg active ingredient).}
\item{REACTPSTmg}{Loss of pesticide from water by reaction during time step (mg active ingredient).}
\item{VOLPSTmg}{Loss of pesticide from water by volatilization during time step (mg active ingredient).}
\item{SETTLPSTmg}{Transfer of pesticide from water to river bed sediment by settling during time step (mg active ingredient).}
\item{RESUSP_PSTmg}{Transfer of pesticide from river bed sediment to water by resuspension during time step (mg active ingredient).}
\item{DIFFUSEPSTmg}{Transfer of pesticide from water to river bed sediment by diffusion during time step (mg active ingredient).}
\item{REACBEDPSTmg}{Loss of pesticide from river bed sediment by reaction during time step (mg active ingredient).}
\item{BURYPSTmg}{Loss of pesticide from river bed sediment by burial during time step (mg active ingredient).}
\item{BED_PSTmg}{Pesticide in river bed sediment during time step (mg active ingredient).}
\item{BACTP_OUTct}{Number of persistent bacteria transported out of reach during time step (# cfu/100 mL).}
\item{BACTLP_OUTct}{Number of less persistent bacteria transported out of reach during time step (# cfu/100 mL).}
\item{CMETAL_1kg}{Conservative metal #1 transported out of reach (kg).}
\item{CMETAL_2kg}{Conservative metal #2 transported out of reach (kg).}
\item{CMETAL_3kg}{Conservative metal #3 transported out of reach (kg).}
\item{TOTNkg}{Total nitrogen transported with water out of reach during time step (kg N).}
\item{TOTPkg}{Total phosphorus transported with water out of reach during time step (kg N).}
\item{NO3_mg_l}{Nitrate concentration in reach N mg/L}
\item{WTMPdegc}{Water temperature degrees C.}
\item{RCH2}{Reach number dublicate.}
}

For further details, see \url{https://swat.tamu.edu/media/116601/ch32_output.pdf}
}
\usage{
daily_rch
}
\description{
Data sample imported from daily SWAT output.rch file with read_swat_output function from data-raw/daily/output.rch.
}
\keyword{datasets}
