% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_averaging.R
\name{get_averaged_collapsed_data}
\alias{get_averaged_collapsed_data}
\title{Function to get averaged by periods and collapsed data by setups}
\usage{
get_averaged_collapsed_data(df, period_list)
}
\arguments{
\item{df}{Data.frame of imported output.*** SWAT file}

\item{period_list}{List of periods used in averaging data  (example
period_list <- list (base = c(2000, 2019), mid = c(2040, 2059), end = c(2080, 2099)))}
}
\value{
Data.frame of averaged data over periods and collapsed to setups
(it means removing reach information).
}
\description{
This function averages output data by periods and collapse it by setups.
}
