% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_averaging.R
\name{get_collapsed_results_to_setups}
\alias{get_collapsed_results_to_setups}
\title{Function to sum up output data over setups}
\usage{
get_collapsed_results_to_setups(df, rch = FALSE)
}
\arguments{
\item{df}{Data.frame of imported output.*** SWAT file}

\item{rch}{TRUE or FALSE to leave reach numbers for subbasin data}
}
\value{
Data.frame of summed up data over setups
}
\description{
This function averages output data imported from SWAT output.rch
and output.sub files. Data should have been imported by using
extract_watersheds_output function.
}
