% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_averaging.R
\name{get_main_parameters}
\alias{get_main_parameters}
\title{Function to sum parameters to NT, PT, ST and Q}
\usage{
get_main_parameters(df)
}
\arguments{
\item{df}{Data.frame of imported output.*** SWAT file}
}
\value{
Data.frame of summed parameters
}
\description{
This function sums output data to main parameters. NT - total nitrogen,
PT - total phosphorus, ST - total sediment and Q - water discharge.
}
