% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_averaging.R
\name{get_period_means}
\alias{get_period_means}
\title{Function to average output data over some period}
\usage{
get_period_means(df, starting_year, ending_year, data_type = "y")
}
\arguments{
\item{df}{Data.frame of imported output.*** SWAT file}

\item{starting_year}{Number representing starting year of period (example 2000)}

\item{ending_year}{Number representing ending year of period (example 2019)}

\item{data_type}{One letter sting "y" for yearly or "m" for monthly}
}
\value{
Data.frame of averaged data over period
}
\description{
This function averages output data imported from SWAT output.rch
output.sub and output.hru files. Data should have been imported
by using extract_watersheds_output function.
}
