% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_averaging.R
\name{get_scenario_sub_sums}
\alias{get_scenario_sub_sums}
\title{Function to get sums of 'sub' data over scenarios}
\usage{
get_scenario_sub_sums(df, period_list)
}
\arguments{
\item{df}{Data.frame of imported output.*** SWAT file}

\item{period_list}{List of periods used in averaging data  (example
period_list <- list (base = c(2000, 2019), mid = c(2040, 2059), end = c(2080, 2099)))}
}
\value{
Data.frame of summed up over scenarios.
}
\description{
This function averages output data by periods and collapse it by setups,
and sums it over scenarios for all setups.
}
