% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monthly_sub}
\alias{monthly_sub}
\title{Monthly output.sub data}
\format{
A data frame with 31 variables:
\describe{
\item{date}{Date}
\item{Year}{Year}
\item{Month}{Month}
\item{Day}{Day}
\item{SUB}{File type}
\item{GIS}{Subbasin number}
\item{MON}{Month}
\item{AREAkm2}{Area of the subbasin (km2).}
\item{PRECIPmm}{Total amount of precipitation falling on the subbasin during time step (mm H2O)."}
\item{SNOMELTmm}{Total amount of snow melt on the subbasin during time step (mm H2O)."}
\item{PETmm}{Potential evapotranspiration from the subbasin during the time step (mm H2O)."}
\item{ETmm}{Actual evapotranspiration from the subbasin during the time step (mm).}
\item{SWmm}{Soil water content (mm). Amount of water in the soil profile at the end of the time period.}
\item{PERCmm}{Water that percolates past the root zone during the time step (mm).}
\item{SURQmm}{Surface runoff contribution to streamflow during time step (mm H2O).}
\item{GW_Qmm}{Groundwater contribution to streamflow (mm). Water from the shallow aquifer that returns to the reach during the time step.}
\item{WYLDmm}{Water yield (mm H2O). The net amount of water that leaves the subbasin and contributes to streamflow in the reach during the time step. (WYLD = SURQ + LATQ + GWQ – TLOSS – pond abstractions).}
\item{SYLDt/ha}{Sediment yield (metric tons/ha). Sediment from the subbasin that is transported into the reach during the time step.}
\item{ORGNkg/ha}{Organic N yield (kg N/ha). Organic nitrogen transported out of the subbasin and into the reach during the time step.}
\item{ORGPkg/ha}{Organic P yield (kg P/ha). Organic phosphorus transported with sediment into the reach during the time step.}
\item{NSURQkg/ha}{NO3 in surface runoff (kg N/ha). Nitrate transported by the surface runoff into the reach during the time step.}
\item{SOLPkg/ha}{Soluble P yield (kg P/ha). Phosphorus that is transported by surface runoff into the reach during the time step.}
\item{SEDPkg/ha}{Mineral P yield (kg P/ha). Mineral phosphorus attached to sediment that is transported by surface runoff into the reach during the time step.}
\item{LATQmm}{Lateral flow contribution to streamflow in watershed for the day, month or year (mm).}
\item{LATNO3kg/ha}{NO3-N in lateral flow (kg N/ha).}
\item{GWNO3kg/ha}{NO3-N in ground flow (kg N/ha).}
\item{CHOLAmic/L}{Concentration of chlorophyll-a for the day, month or year (mic/L).}
\item{CBODUmg/L}{Carbonaceous biochemical oxygen demand of material transported during time step (mg O2/L).}
\item{DOXQmg/L}{Amount of dissolved oxygen transported during time step (mg/L).}
\item{TNO3kg/ha}{NO3-N in tile drain flow (kg N/ha).}
\item{RCH}{Reach number for the subbasin.}
}

For further details, see \url{https://swat.tamu.edu/media/116601/ch32_output.pdf}
}
\usage{
monthly_sub
}
\description{
Data sample imported from monthly SWAT output.sub file with read_swat_output function from data-raw/monthly/output.sub.
}
\keyword{datasets}
