% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plotting_results.R
\name{plot_sc_grid_bars}
\alias{plot_sc_grid_bars}
\title{Function to plot grid of bars plots comparing scenarios of measures during different periods.}
\usage{
plot_sc_grid_bars(df, param, rcp)
}
\arguments{
\item{df}{Data.frame of imported output.*** SWAT file. Data
should be summed to one per scenario and period and showing relative change to
baseline.}

\item{param}{String showing, which parameters of df to put on a plot.}

\item{rcp}{String showing, which climate change scenario to use (currently only "rcp45" or "rcp85" available)}
}
\value{
Grid of bar plots
}
\description{
This function put out modeling results on a grid of bars plots.
This grid plot of plots allows to see effectiveness of measures during different time periods.
Also allows to compare results of different RCM.
}
\examples{
##plot_sc_grid_bars(df, "NT", "rcp45")
}
