% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_plotting_results.R
\name{plot_scenario_result_map}
\alias{plot_scenario_result_map}
\title{Function to plot on map modeling results of scenario on basins polygons}
\usage{
plot_scenario_result_map(df, pattern, parameter)
}
\arguments{
\item{df}{Data.frame of imported output.*** SWAT file. Data
should be summed up/averaged to only one per scenario per reach or setup.}

\item{pattern}{String neccessary to filter out one particural scenario.}

\item{parameter}{String showing, which parameters of df to put on a map.}
}
\value{
Map for modeling results on basins polygons
}
\description{
This function put out modeling results on basins polygons used
in modeling. However only one scenario could be plotted. Also data input
should be summed up/averaged to one value for reach or setup/
}
\examples{
##plot_scenario_result_map(df, "RACMO22E_rcp45_measure_33_2040_2059", "NT")
}
