% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_results.R
\name{plot_scenarios_bars}
\alias{plot_scenarios_bars}
\title{Function plot bar plots to compare results from scenarios}
\usage{
plot_scenarios_bars(df, variable_name, filter_vec = c("_"))
}
\arguments{
\item{df}{Data.frame of data extracted from output.rch and output.sub files and processed with get_diff_from_baseline function.
Beside other columns has to have SCENARIO column, which join all information about scenario.}

\item{variable_name}{Sting with the name of column to plot (example "NSURQt/y")}

\item{filter_vec}{Vector, which allows to filter scenarios. Filtering could be done by AND providing
vector members seperated by comma, or by OR seperating one vector member in many '|'. For example
c("measure_40", "rcp45|rcp85", "2080_2099") this vector filter by AND and by OR.}
}
\value{
Bar plot
}
\description{
This function allows plotting bar plots for the scenario results extracted and averaged
from SWAT output files.
}
\examples{
##plot_scenarios_bars(df, "NSURQt/y", c("measure_40", "rcp45|rcp85", "2080_2099"))
}
