% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output.R
\name{read_swat_output}
\alias{read_swat_output}
\title{Function to read directly SWAT model output files from folder paths}
\usage{
read_swat_output(folder, output_type, output_file, starting_date, ending_date)
}
\arguments{
\item{folder}{String giving path to the output.*** folder. (example "data-raw/monthly")}

\item{output_type}{Number the same as in IPRINT (0 - Monthly, 1 - Daily, 2 - Year)}

\item{output_file}{String showing ending of output file. Currently implemented are "rch", "sub", "hru" files.}

\item{starting_date}{Sting with time series starting date (example "1997-01-01")}

\item{ending_date}{Sting with time series ending date (example "2019-12-31")}
}
\value{
Data.frame imported from SWAT output files
}
\description{
This function allows reading all output.rch, output.sub, output.hru files generated by SWAT2012
}
