% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rivers}
\alias{rivers}
\title{Rivers data}
\format{
A sf data.frame with 2 variables:
\describe{
\item{CID}{Catchment ID}
\item{geometry}{Geometry of objects}
}
}
\usage{
rivers
}
\description{
Data of river network used in modeling.
}
\keyword{datasets}
