% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_extract_watersheds_output.R
\name{extract_watersheds_output}
\alias{extract_watersheds_output}
\title{Function to extract all output data from SWAT model output.rch, output.sub and output.hru files}
\usage{
extract_watersheds_output(
  watersheds_folder_path,
  scenarios,
  output_type,
  starting_date,
  ending_date,
  output_files = c("sub", "rch", "hru"),
  save_results_to_files = FALSE
)
}
\arguments{
\item{watersheds_folder_path}{String giving path to the Watersheds folder where all output results are saved.
(example "../../MODEL/MODEL_CLIMATE_LONG/Setup/Watersheds")}

\item{scenarios}{Vector providing scenario name used in saving modeling results. Example for combining climate and measures could be:
as.vector(outer(c("C1", "C2"), c("M1", "M2"), paste, sep="_")). This would allow generation of vector, which could be used in loop.
An example, scenarios = c("HIRHAM5_rcp45_measure_30'", "HIRHAM5_rcp45_measure_31", "HIRHAM5_rcp85_measure_30", "HIRHAM5_rcp85_measure_31")}

\item{output_type}{Number the same as in IPRINT (0 - Monthly, 1 - Daily, 2 - Year)}

\item{starting_date}{Sting with time series starting date (example "1997-01-01")}

\item{ending_date}{Sting with time series ending date (example "2019-12-31")}

\item{output_files}{Vector of files endings for output files to extract.}

\item{save_results_to_files}{Logical value with TRUE or FALSE values. If TRUE  results will be saved to hru.rda,
rch.rda, sub.rda files in working directory.}
}
\value{
List with three data.frames: rch, sub and hru for data extracted from different output files.
}
\description{
This function allows reading in all output.rch, output.sub, output.hru files generated by SWAT2012
in Lithuanian EPA model, which was applied for different climate and measures scenarios.
}
\examples{
r <- extract_watersheds_output("/Watersheds", c("C1", "C2"), 2, "1997-01-01", "2099-12-31")
}
